// web/platform-admin-web/src/api/generated/model/updateOrganizationRoleRequestDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

export interface UpdateOrganizationRoleRequestDto {
  /** 조직 역할 식별자 */
  id: string;
  /** 조직 역할 코드 */
  code?: string;
  /** 조직 역할 이름 */
  name?: string;
  /** 조직 역할 설명 */
  description?: string;
}