// web/platform-admin-web/src/api/generated/model/updateOrganizationRoleSortRequestDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */
import type { UpdateOrganizationRoleSortItemDto } from './updateOrganizationRoleSortItemDto';

export interface UpdateOrganizationRoleSortRequestDto {
  /** 정렬 대상 목록 */
  items: UpdateOrganizationRoleSortItemDto[];
}