// web/platform-admin-web/src/api/generated/model/updatePermissionSetPermissionsRequestDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

export interface UpdatePermissionSetPermissionsRequestDto {
  /** 권한 세트 식별자 */
  id: string;
  /** 권한 코드 목록 */
  permissionCodes?: string[];
}